the project returned to its original state, with cron updates

+++ src/office/office.service.ts
@@ -19,7 +19,6 @@ import { UpdateOfficeStatusDto } from './dto/update-office-status.dto';
 import { PaginationDto } from 'src/common/utils/pagination.dto';
 import { PaginatedResponse } from 'src/common/utils/paginated-response.interface';
 import { OfficeSubscription } from 'src/office-subscription/entities/office-subscription.entity';
-import { of } from 'rxjs';
 
 @Injectable()
 export class OfficeService {
@@ -43,7 +42,7 @@ export class OfficeService {
    */
   async create(
     createOfficeDto: CreateOfficeDto,
-    /*userId: string,*/ /**omar */officeId: string,
+    userId: string,
     license_photo?: Express.Multer.File,
     office_photo?: Express.Multer.File,
   ) {
@@ -52,83 +51,50 @@ export class OfficeService {
     await queryRunner.startTransaction();
 
     try {
-      // if (!license_photo || !office_photo) {
-      //   throw new BadRequestException(
-      //     'license_photo or office_photo is missed',
-      //   );
-      //}
+      if (!license_photo || !office_photo) {
+        throw new BadRequestException(
+          'license_photo or office_photo is missed',
+        );
+      }
 
-      //omar comment this
-      // const user = await this.userRepository.findOne({
-      //   where: { id: userId },
-      // });
-      //if (!user) throw new NotFoundException('User not found');
+      const user = await this.userRepository.findOne({
+        where: { id: userId },
+      });
+      if (!user) throw new NotFoundException('User not found');
 
       let office = await this.officeRepository.findOne({
         where: {
-          //user,
-
-          //omar
-          id: officeId
+          user,
         },
       });
 
-      console.log(office?.office_email)
-      console.log(officeId)
-            console.log(office);
-
-
-      // if (office) {
-      //   throw new BadRequestException("you can't have multiple offices");
-      // }
-
-      var licensePhoto;
-      var officePhotoEntity;
-
-      if(license_photo){
-
-        const uploadedLicense =
-          await this.cloudinaryService.uploadImage(license_photo);
-
-        licensePhoto = this.licensePhotoRepository.create({
-          url: uploadedLicense.secure_url,
-          public_id: uploadedLicense.public_id,
-        });
-
-        await queryRunner.manager.save(licensePhoto);
-
+      if (office) {
+        throw new BadRequestException("you can't have multiple offices");
       }
+      const uploadedLicense =
+        await this.cloudinaryService.uploadImage(license_photo);
+      const uploadedOffice =
+        await this.cloudinaryService.uploadImage(office_photo);
 
-      if(office_photo){
+      const licensePhoto = this.licensePhotoRepository.create({
+        url: uploadedLicense.secure_url,
+        public_id: uploadedLicense.public_id,
+      });
 
-        const uploadedOffice =
-          await this.cloudinaryService.uploadImage(office_photo);
+      const officePhotoEntity = this.officePhotoRepository.create({
+        url: uploadedOffice.secure_url,
+        public_id: uploadedOffice.public_id,
+      });
 
-        officePhotoEntity = this.officePhotoRepository.create({
-          url: uploadedOffice.secure_url,
-          public_id: uploadedOffice.public_id,
-        });
+      await queryRunner.manager.save(licensePhoto);
+      await queryRunner.manager.save(officePhotoEntity);
 
-        await queryRunner.manager.save(officePhotoEntity);
-
-      }
-
-      if(office){
-        office.license_number = createOfficeDto.license_number ;
-        office.password = createOfficeDto.password;
-        office.name = createOfficeDto.name;
-        office.personal_identity_number = createOfficeDto.personal_identity_number;
-        office.receiver_identifier = createOfficeDto.receiver_identifier;
-
-        if(licensePhoto){
-          office.license_photo = licensePhoto;
-        }
-
-        if(officePhotoEntity){
-          office.office_photo = officePhotoEntity
-        }
-
-      }      
+      office = this.officeRepository.create({
+        ...createOfficeDto,
+        user,
+        license_photo: licensePhoto,
+        office_photo: officePhotoEntity,
+      });
 
       await queryRunner.manager.save(office);
       await queryRunner.commitTransaction();
@@ -361,10 +327,7 @@ export class OfficeService {
   async update(
     officeId: string,
     updateOfficeDto: UpdateOfficeDto,
-
-    //omar comment this
-    /*userId: string,*/ 
-
+    userId: string,
     office_photo?: Express.Multer.File,
   ) {
     const queryRunner = this.dataSource.createQueryRunner();
@@ -390,10 +353,9 @@ export class OfficeService {
 
       if (!office) throw new NotFoundException('Office not found');
 
-      //omar comment this
-      // if (office.user.id !== userId) {
-      //   throw new ForbiddenException("it's forbidden to do this!");
-      // }
+      if (office.user.id !== userId) {
+        throw new ForbiddenException("it's forbidden to do this!");
+      }
 
       if (office_photo) {
         if (office.office_photo?.public_id) {
@@ -479,10 +441,9 @@ export class OfficeService {
 
       if (!office) throw new NotFoundException('Office not found');
 
-      //omar comment this
-      // if (office.user.id !== userId) {
-      //   throw new ForbiddenException("it's forbidden to do this!");
-      // }
+      if (office.user.id !== userId) {
+        throw new ForbiddenException("it's forbidden to do this!");
+      }
 
       if (office.license_photo?.public_id) {
         await this.cloudinaryService.deleteImage(
